refactor: extract MCP connection logic to helper module
- Move MCP server connection logic to registerTestMcp.ts helper
- Clean up agent-conversation.ts by extracting MCP setup
- Keep MCP connections (Figma SSE + Linear stdio) in one place
- Maintain separation of concerns for better maintainability
